refactor: view 패키지를 was로 이동
